Final Implementation 1.0/backend/requirement.txt
Flask==3.0.2
Flask-Cors==4.0.0
praw==7.7.1
joblib==1.3.2
pandas==2.2.1
pymongo==4.6.2
python-dotenv==1.0.1
google-generativeai==0.3.2
scikit-learn==1.6.1
